requirements.txt
aiohttp>=3.11.18
uuid6>=2024.7.10
Pillow>=10.4.0
filetype>=1.2.0
